+++ inference/python_api_test/test_ocr_model/run_ce_jetson.sh
